Revert "Fix issue when int64_t was an unknown type"
This reverts commit 5ce77c5bc0f1008cc7d60960f9f9c27f3230c671.

+++ LuaCppInterface/luatable.h
@@ -5,8 +5,6 @@
 #include <string>
 #include <functional>
 #include <memory>
-#include <stdint.h>
-
 #include "luareference.h"
 
 template<typename SIG>
@@ -75,7 +73,7 @@ public:
 
 	// Get all the keys in the LuaTable
 	void ForAllStringKeys(std::function<void(std::string, LuaType::Value)> stringKeys) const;
-	void ForAllIntegerKeys(std::function<void(size_t, LuaType::Value)> integerKeys) const;
+	void ForAllIntegerKeys(std::function<void(int64_t, LuaType::Value)> integerKeys) const;
 
 };
